// src/lib/db-types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Albums {
  author_id: number;
  id: Generated<number>;
  name: string;
  release_date: number;
}

export interface Authors {
  bio: string | null;
  id: Generated<number>;
  name: string;
}

export interface Songs {
  album_id: number;
  duration: number;
  id: Generated<number>;
  name: string;
  release_date: number;
}

export interface DB {
  albums: Albums;
  authors: Authors;
  songs: Songs;
}